// src/types/discover/discover.ts
export interface Movie {
	id: string;
	title: string;
	poster_path: string;
	adult: boolean;
	backdrop_path: string;
	belongs_to_collection?: {
		id: string;
		name: string;
		poster_path: string;
		backdrop_path: string;
	}[];
	budget: number;
	genres: {
		id: string;
		name: string;
	}[];
	homepage: string;
	imdb_id: string;
	origin_country: string[];
	original_language: string;
	original_title: string;
	overview: string;
	popularity: number;
	production_companies: {
		id: string;
		logo_path: string;
		name: string;
		origin_country: string;
	}[];
	production_countries: {
		iso_3166_1: string;
		name: string;
	}[];
	release_date: string;
	revenue: number;
	runtime: number;
	spoken_languages: {
		english_name: string;
		iso_639_1: string;
		name: string;
	}[];
	status: string;
	tagline: string;
	video: boolean;
	vote_average: number;
	vote_count: number;
}

export interface FetchTrendingMoviesResponse {
	page: number;
	results: Movie[];
	total_pages: number;
	total_results: number;
}

export interface FetchTrendingMoviesQueryResult {
	pageParams: number[];
	pages: FetchTrendingMoviesResponse[];
}